Eliminated use of Trace. Eliminated use of TestThreadObjectContainer. Use of global container is constrained to the Publisher and only at start-up. Eliminated unused Location property from FeatureLevelCucumberMessages; Eliminated CucumberMessageFactoryInner;
