+++ Reqnroll/Formatters/RuntimeSupport/FormatterLog.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Reqnroll.Formatters.RuntimeSupport
+{
+    public class FormatterLog : IFormatterLog
+    {
+        public void WriteMessage(string message)
+        {
+            Debug.WriteLine(message);
+        }
+    }
+}
