+++ Reqnroll/Formatters/RuntimeSupport/IFormatterLog.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Reqnroll.Formatters.RuntimeSupport
+{
+    public interface IFormatterLog
+    {
+        public void WriteMessage(string message);
+    }
+}
