+++ Tests/Reqnroll.RuntimeTests/Formatters/ExecutionTracking/FeatureExecutionTrackerTests.cs
@@ -58,8 +58,7 @@ namespace Reqnroll.RuntimeTests.Formatters.ExecutionTracking
             var featureLevelCucumberMessagesDummy = new FeatureLevelCucumberMessages(
                 () => new Source("c:\\file", "Feature test", SourceMediaType.TEXT_X_CUCUMBER_GHERKIN_PLAIN),
                 () => new GherkinDocument("", new Feature(new Location(0, 0), [], "en", "Feature", "Dummy Feature", "", new List<FeatureChild>()), []),
-                () => new List<Pickle>([new Pickle("0", "", "", "en", [new PickleStep(null, [], "step1", PickleStepType.ACTION, "I eat a cuke")], [], [])]),
-                "location0");
+                () => new List<Pickle>([new Pickle("0", "", "", "en", [new PickleStep(null, [], "step1", PickleStepType.ACTION, "I eat a cuke")], [], [])]));
 
             _featureInfoDummy = new FeatureInfo(CultureInfo.InvariantCulture, null, "Test Feature", null, ProgrammingLanguage.CSharp, null, featureLevelCucumberMessagesDummy);
             mockFeatureContext.Setup(m => m.FeatureInfo).Returns(_featureInfoDummy);
